Use common component for IRC27
